+++ Reinvest/Registration/src/Adapter/Database/Repository/NorthCapitalDocumentsSynchronizationRepository.ts
@@ -1,15 +1,18 @@
 import {IdGeneratorInterface} from "IdGenerator/IdGenerator";
 import {
     northCapitalDocumentsSynchronizationTable,
-    RegistrationDatabaseAdapterProvider,
+    RegistrationDatabaseAdapterProvider, registrationMappingRegistryTable,
 } from "Registration/Adapter/Database/DatabaseAdapter";
 import {
     InsertableNorthCapitalDocumentsSynchronization
 } from "Registration/Adapter/Database/RegistrationSchema";
 import {
+    DocumentSyncState,
+    NorthCapitalDocumentToSync,
     NorthCapitalObjectType,
 } from "Registration/Domain/VendorModel/NorthCapital/NorthCapitalTypes";
 import {DocumentSchema} from "Registration/Domain/Model/ReinvestTypes";
+import {MappedRecordStatus} from "Registration/Domain/Model/Mapping/MappedType";
 
 export class NorthCapitalDocumentsSynchronizationRepository {
     public static getClassName = (): string => "NorthCapitalDocumentsSynchronizationRepository";
@@ -22,10 +25,8 @@ export class NorthCapitalDocumentsSynchronizationRepository {
         this.idGenerator = uniqueGenerator;
     }
 
-
     async addDocuments(recordId: string, northCapitalId: string, northCapitalType: NorthCapitalObjectType, documents: DocumentSchema[]): Promise<void> {
         const values = documents.map((document: DocumentSchema) => (<InsertableNorthCapitalDocumentsSynchronization>{
-            id: this.idGenerator.createUuid(),
             recordId,
             northCapitalId,
             northCapitalType,
@@ -53,8 +54,8 @@ export class NorthCapitalDocumentsSynchronizationRepository {
             const data = await this.databaseAdapterProvider.provide()
                 .selectFrom(northCapitalDocumentsSynchronizationTable)
                 .select(['documentId'])
-                .where('state', '=', 'DIRTY')
-                .orWhere('state', '=', 'TO_BE_DELETED')
+                .where('state', '=', DocumentSyncState.DIRTY)
+                .orWhere('state', '=', DocumentSyncState.TO_BE_DELETED)
                 .orderBy('updatedDate', 'asc')
                 .limit(20)
                 .execute();
@@ -65,5 +66,61 @@ export class NorthCapitalDocumentsSynchronizationRepository {
             return [];
         }
 
+    }
+
+    async getDocumentToSync(documentId: string): Promise<NorthCapitalDocumentToSync | never> {
+        return await this.databaseAdapterProvider.provide()
+            .selectFrom(northCapitalDocumentsSynchronizationTable)
+            .select(['recordId', 'northCapitalId', 'northCapitalType', 'documentId', 'documentPath', 'documentFilename', 'version', 'state', 'createdDate', 'updatedDate'])
+            .where('documentId', '=', documentId)
+            .limit(1)
+            .castTo<NorthCapitalDocumentToSync>()
+            .executeTakeFirstOrThrow();
+    }
+
+    async setClean(document: NorthCapitalDocumentToSync): Promise<boolean> {
+        const updatedDate = new Date();
+        const currentVersion = document.version;
+
+        try {
+            await this.databaseAdapterProvider.provide()
+                .updateTable(northCapitalDocumentsSynchronizationTable)
+                .set({
+                    state: DocumentSyncState.CLEAN,
+                    version: currentVersion + 1,
+                    updatedDate,
+                })
+                .where('documentId', '=', document.documentId)
+                .where('version', '=', currentVersion)
+                .returning(['version'])
+                .executeTakeFirstOrThrow();
+
+            return true;
+        } catch (error: any) {
+            console.error('Failed to set document as clean - probably race conditioning', error.message);
+            return false;
+        }
+    }
+
+    async setFailed(document: NorthCapitalDocumentToSync): Promise<void> {
+        const updatedDate = new Date();
+        const currentVersion = document.version;
+
+        try {
+            await this.databaseAdapterProvider.provide()
+                .updateTable(northCapitalDocumentsSynchronizationTable)
+                .set({
+                    state: DocumentSyncState.FAILED,
+                    version: currentVersion + 1,
+                    updatedDate,
+                })
+                .where('documentId', '=', document.documentId)
+                .where('version', '=', currentVersion)
+                .returning(['version'])
+                .executeTakeFirstOrThrow();
+
+        } catch (error: any) {
+            console.error('Failed to set document as failed - probably race conditioning', error.message);
+        }
     }
 }