+++ Reinvest/Registration/src/Port/Api/RegistrationApiType.ts
@@ -1,10 +1,15 @@
 import {ContainerInterface} from "Container/Container";
-import {NorthCapitalDocumentSynchronizationQuery} from "Registration/Port/Api/DocumentSynchronizationQuery";
+import {
+    NorthCapitalDocumentSynchronizationController
+} from "Registration/Port/Api/NorthCapitalDocumentSynchronizationController";
+import {NorthCapitalDocumentSynchronizationQuery} from "./NorthCapitalDocumentSynchronizationQuery";
 
 export type RegistrationApiType = {
-    listDocumentsToSynchronize: NorthCapitalDocumentSynchronizationQuery['listDocumentsToSynchronize'];
+    listDocumentsToSynchronize: NorthCapitalDocumentSynchronizationQuery['listDocumentsToSynchronize'],
+    synchronizeDocument: NorthCapitalDocumentSynchronizationController['synchronizeDocument'],
 }
 
 export const registrationApi = (container: ContainerInterface): RegistrationApiType => ({
-    listDocumentsToSynchronize: container.delegateTo(NorthCapitalDocumentSynchronizationQuery, 'listDocumentsToSynchronize')
+    listDocumentsToSynchronize: container.delegateTo(NorthCapitalDocumentSynchronizationQuery, 'listDocumentsToSynchronize'),
+    synchronizeDocument: container.delegateTo(NorthCapitalDocumentSynchronizationController, 'synchronizeDocument')
 });